Switch theme for microsite group view page to admin theme.
Working with a theme switcher rather than adding _admin to the route, in
the case that there might be different types of groups.

Plan to add to domain_group the logic that sets if a group can have a
domain, so the class can be applied to any group defined as a domain
group. For now it's hard coded to microsite group type.