Restore views #27
These will override the group views if installed together. If group is
intalled first the existing views are used.

Hard coded to override the view for just the microsite group type.
Assuming no other group types.

This wants further work:
 * Allow for additional microsite group types
 * Allow for non microsite group types.

Probably need a couple of hooks to decide which group types there are
and if the view is not otherwised changed update its settings to display
the view pages as tabs on the appropriate group types.
Can't think of a better way round that at the moment, bit fiddly though,
especially deciding what state the view should be in before adding to
it.