Inital fix to replicate exisitng content into a group as created.
Notes:-

Uses config localgov_microsites_group.settings[default_group_node]
which is not set, nor is there a UI to set it, defaults to node 1.
Follow-up: Add UI, default value in profile to the imported default
content.

Replication of entity with paragraphs will always cause them to be
cloned, no other referenced content will be cloned. Questinon (and todo)
in the code to decide if this should be restricted (by some value on the
entity).