Add media uploaded via the library form in a group domain to the group.
https://github.com/localgovdrupal/localgov_microsites_group/issues/348#issuecomment-1330991579

The first option explored. Add the a hook so that groupmedia would know
the group based on domain; but groupmedia is only adding a group
relation to existing media that is then attached to another entity that
is in a group.

Second option then. When the media is inserted add the relation. However
there's no way of knowing if it's being added using the group create
form [eg /group/1/content/create/group_media%3Aimage] in which case you
end up duplicating, and more, the relations.

So finally target the specific forms where media is uploaded and isn't
placed into a group.