Also override the defalt view display.
Question: Is this wise?

If the page view in the group module changes upstream, with assumptions
about the default this could break.

Seen the other way it stops the deault breaking our view.

Alternative? Override the other parts of the view that are being added
in this MR in our displays, as has been done so far.