Add field_domain_access value to domain_path forms #338
The Domain Path module does a lot of checks for domain access configuration, but we use the Group module for access checks. Adding the domain access field to the form ensures the Domain Path checks pass, but is a bit of a hack as it would be better to fix whatever is going wrong in the Domain Path module.